Move ApiJsonResponses to concern
